+++ CHANGELOG.md
@@ -6,8 +6,9 @@ See [GitHub releases](https://github.com/pyOpenSci/pyosMeta/releases) page for a
 
 ## [Unreleased]
 
-* Feat: add tqdm progress bar to parse issues workflow & remove community heading noise (@lwasser, #385)
-* Fix: permissions issue in reusable workflow
+* Fix: handle GitHub 401 and 403 API errors, use REST API for metadata, ensure metadata are updated, and fail on bad API calls (@lwasser, #384)
+* Fix: permissions issue in reusable workflow (@lwasser, #382)
+* Feat: add `tqdm` progress bar to parse issues workflow & remove community heading noise (@lwasser, #385)
 
 ## [v1.7.9] - 2026-08-05